requirements.txt
# Azure Services
azure-ai-formrecognizer==3.3.0
azure-core==1.30.1

# Deepgram
deepgram-sdk==3.2.7

# Google Gemini
google-generativeai==0.5.4

# API Clients
requests==2.31.0
arxiv==2.1.0
biopython==1.83  # For PubMed API

# NLP and ML
langchain==0.1.16
langchain-community==0.0.34
chromadb==0.4.24
sentence-transformers==2.6.1
openai==1.23.2

# Utilities
python-dotenv==1.0.1
pydantic==2.7.0
tenacity==8.2.3
tqdm==4.66.2

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Logging
loguru==0.7.2